requirements.txt
numpy==1.23.0
argbind>=0.3.7
descript-audiotools>=0.7.2
descript-audio-codec
einops
torchaudio
tqdm
tensorboard
numba>=0.5.7
pandas
openpyxl
jupyterlab


